+++ ferredesk_v0/backend/ferreapps/ventas/views.py
@@ -1,1842 +1,49 @@
-from django.shortcuts import render, get_object_or_404
-from rest_framework import viewsets, status
-from rest_framework.decorators import action, api_view, permission_classes
-from rest_framework.response import Response
-from django.http import HttpResponse, Http404
-from django.core.exceptions import ValidationError
-from django.db.models import ProtectedError
-from .models import Comprobante, Venta, VentaDetalleItem, VentaDetalleMan, VentaRemPed, VentaDetalleItemCalculado, VentaIVAAlicuota, VentaCalculada
-from .serializers import (
-    ComprobanteSerializer,
-    VentaSerializer,
-    VentaDetalleItemSerializer,
-    VentaDetalleManSerializer,
-    VentaRemPedSerializer,
-    VentaDetalleItemCalculadoSerializer,
-    VentaIVAAlicuotaSerializer,
-    VentaCalculadaSerializer
+"""
+DEPRECATED: Este archivo se mantiene por compatibilidad.
+Todo el código se ha movido a la carpeta views/
+
+Importar desde: from ferreapps.ventas.views import ...
+"""
+import warnings
+
+# Re-exportar todo desde el nuevo módulo views/
+from .views import (
+    # Utilidades stock
+    _obtener_stock_proveedores_bloqueado,
+    _total_disponible_en_proveedores,
+    _obtener_codigo_venta,
+    _obtener_nombre_proveedor,
+    _obtener_proveedor_habitual_stock,
+    _descontar_distribuyendo,
+    # Comprobantes
+    ComprobanteViewSet,
+    # Ventas
+    VentaFilter,
+    VentaCalculadaFilter,
+    VentaViewSet,
+    VentaDetalleItemViewSet,
+    VentaDetalleManViewSet,
+    VentaRemPedViewSet,
+    VentaDetalleItemCalculadoFilter,
+    VentaDetalleItemCalculadoViewSet,
+    VentaIVAAlicuotaFilter,
+    VentaIVAAlicuotaViewSet,
+    VentaCalculadaViewSet,
+    # Conversiones
+    convertir_presupuesto_a_venta,
+    _eliminar_auto_imputaciones_cliente_generico,
+    convertir_factura_interna_a_fiscal,
+    verificar_imputaciones_comprobante,
+    # Dashboard
+    productos_mas_vendidos,
+    ventas_por_dia,
+    clientes_mas_ventas,
 )
-from django.db import transaction
-from ferreapps.productos.models import Stock, StockProve, Proveedor
-from decimal import Decimal
-from ferreapps.productos.models import Ferreteria
-from ferreapps.clientes.models import Cliente, TipoIVA
-from django.db import IntegrityError
-import logging
-from rest_framework.permissions import IsAuthenticated
-from .ARCA.settings_arca import COMPROBANTES_INTERNOS
 
-# Configurar logger para este módulo
-logger = logging.getLogger(__name__)
-from .utils import asignar_comprobante, _construir_respuesta_comprobante
-from django_filters.rest_framework import DjangoFilterBackend, FilterSet, DateFromToRangeFilter, NumberFilter, CharFilter
-
-# Punto de venta fijo para comprobantes internos/no oficiales (se formatea como 0099 en UI)
-PUNTO_VENTA_INTERNO = 99
-
-# ------------------------------------------------------------
-# Utilidades internas para gestión de stock entre proveedores
-# ------------------------------------------------------------
-def _obtener_stock_proveedores_bloqueado(stock_id):
-    """
-    Devuelve la lista de StockProve del producto (stock_id) con bloqueo select_for_update.
-    """
-    return list(StockProve.objects.select_for_update().filter(stock_id=stock_id))
-
-
-def _total_disponible_en_proveedores(stock_id):
-    """
-    Suma el stock disponible entre todos los proveedores de un producto.
-    Retorna (total: Decimal, proveedores: List[StockProve] bloqueados).
-    """
-    proveedores = _obtener_stock_proveedores_bloqueado(stock_id)
-    total = sum((sp.cantidad for sp in proveedores), Decimal('0'))
-    return total, proveedores
-
-
-def _obtener_codigo_venta(stock_id):
-    """
-    Obtiene el código de venta (STO_CODVTA) del producto asociado al stock_id.
-    Si no se encuentra, devuelve el propio stock_id como fallback.
-    """
-    try:
-        stock = Stock.objects.get(id=stock_id)
-        # El campo del código de venta en el modelo Stock suele ser 'codigo_venta' o similar; 
-        # revisar atributos disponibles y usar el que exista.
-        codigo = getattr(stock, 'codigo_venta', None) or getattr(stock, 'codvta', None) or getattr(stock, 'STO_CODVTA', None)
-        return codigo or str(stock_id)
-    except Stock.DoesNotExist:
-        return str(stock_id)
-
-
-def _obtener_nombre_proveedor(proveedor_id):
-    """
-    Obtiene el nombre del proveedor (razón social o fantasía) asociado al proveedor_id.
-    Si no se encuentra, devuelve el propio proveedor_id como fallback.
-    """
-    try:
-        proveedor = Proveedor.objects.get(id=proveedor_id)
-        return proveedor.razon or proveedor.fantasia or str(proveedor_id)
-    except Proveedor.DoesNotExist:
-        return str(proveedor_id)
-
-
-def _obtener_proveedor_habitual_stock(stock_id):
-    """
-    Obtiene el proveedor habitual de un stock.
-    Como todos los productos tienen proveedor habitual obligatorio, esto nunca debería fallar.
-    """
-    try:
-        stock = Stock.objects.get(id=stock_id)
-        return stock.proveedor_habitual.id
-    except Stock.DoesNotExist:
-        return None
-    except AttributeError:
-        # Si por alguna razón el proveedor_habitual es None (no debería ocurrir)
-        return None
-
-
-def _descontar_distribuyendo(stock_id, proveedor_preferido_id, cantidad, permitir_stock_negativo, errores_stock, stock_actualizado):
-    """
-    Descuenta "cantidad" del stock del producto (stock_id), priorizando el proveedor preferido,
-    y luego el resto de proveedores del mismo producto hasta cubrir la cantidad.
-
-    - Si permitir_stock_negativo es True: primero distribuye todo el stock disponible, 
-      luego descuenta lo que falte del proveedor preferido (puede quedar negativo).
-    - Si permitir_stock_negativo es False: intenta distribuir. Si la suma total no alcanza, 
-      no descuenta y agrega error.
-
-    Agrega una entrada en stock_actualizado por cada proveedor afectado.
-    Retorna True si se aplicó el descuento, False si no fue posible (y se registró el error).
-    """
-    try:
-        cantidad = Decimal(str(cantidad))
-    except Exception:
-        cantidad = Decimal('0')
-
-    # Obtener todos los proveedores con stock bloqueado
-    total_disponible, proveedores_bloqueados = _total_disponible_en_proveedores(stock_id)
-    
-    # Si no permitir stock negativo, validar que alcance
-    if not permitir_stock_negativo:
-        if total_disponible < cantidad:
-            cod = _obtener_codigo_venta(stock_id)
-            errores_stock.append(
-                f"Stock insuficiente para producto {cod}. Disponible total: {total_disponible}, solicitado: {cantidad}"
-            )
-            return False
-    
-    # Mapear por proveedor para acceso rápido (todas las instancias están bloqueadas)
-    prov_map = {sp.proveedor_id: sp for sp in proveedores_bloqueados}
-    orden_proveedores = []
-
-    # 1) Proveedor preferido primero (si existe)
-    if proveedor_preferido_id in prov_map:
-        orden_proveedores.append(proveedor_preferido_id)
-
-    # 2) Resto de proveedores por mayor disponibilidad
-    resto = [sp for sp in proveedores_bloqueados if sp.proveedor_id != proveedor_preferido_id]
-    resto.sort(key=lambda x: x.cantidad, reverse=True)
-    orden_proveedores.extend([sp.proveedor_id for sp in resto])
-
-    # Distribuir primero todo el stock disponible
-    restante = cantidad
-    for prov_id in orden_proveedores:
-        if restante <= 0:
-            break
-        sp = prov_map[prov_id]
-        disponible = Decimal(str(sp.cantidad))
-        if disponible <= 0:
-            continue
-        consumir = min(disponible, restante)
-        sp.cantidad = disponible - consumir
-        sp.save()
-        stock_actualizado.append((sp.stock_id, sp.proveedor_id, sp.cantidad))
-        restante -= consumir
-
-    # Si aún queda cantidad por descontar y está permitido stock negativo
-    if restante > 0 and permitir_stock_negativo:
-        # Verificar que existe el proveedor preferido
-        if proveedor_preferido_id not in prov_map:
-            cod = _obtener_codigo_venta(stock_id)
-            nombre_proveedor = _obtener_nombre_proveedor(proveedor_preferido_id)
-            errores_stock.append(f"No existe stock para el producto {cod} y proveedor {nombre_proveedor}")
-            return False
-        
-        # Descontar lo que falte del proveedor preferido (puede quedar negativo)
-        sp = prov_map[proveedor_preferido_id]
-        sp.cantidad -= restante
-        sp.save()
-        stock_actualizado.append((sp.stock_id, sp.proveedor_id, sp.cantidad))
-        restante = 0
-    
-    # Verificación final de seguridad
-    if restante > 0:
-        # No debería ocurrir, pero por seguridad
-        cod = _obtener_codigo_venta(stock_id)
-        errores_stock.append(
-            f"Stock insuficiente para producto {cod}. Disponible total: {total_disponible}, solicitado: {cantidad}"
-        )
-        return False
-
-    return True
-
-# Importación para integración ARCA automática
-from .ARCA import emitir_arca_automatico, debe_emitir_arca, FerreDeskARCAError
-
-# Diccionario de alícuotas (igual que en el frontend)
-ALICUOTAS = {
-    1: Decimal('0'),  # NO GRAVADO
-    2: Decimal('0'),  # EXENTO
-    3: Decimal('0'),  # 0%
-    4: Decimal('10.5'),
-    5: Decimal('21'),
-    6: Decimal('27')
-}
-
-# Create your views here.
-
-class ComprobanteViewSet(viewsets.ModelViewSet):
-    queryset = Comprobante.objects.all()
-    serializer_class = ComprobanteSerializer
-
-    @action(detail=False, methods=['post'], url_path='asignar')
-    def asignar(self, request):
-        tipo_comprobante = request.data.get('tipo_comprobante')
-        situacion_iva_cliente = request.data.get('situacion_iva_cliente')
-        if not tipo_comprobante or not situacion_iva_cliente:
-            return Response({'detail': 'Faltan datos requeridos.'}, status=status.HTTP_400_BAD_REQUEST)
-        try:
-            comprobante = asignar_comprobante(tipo_comprobante, situacion_iva_cliente)
-            return Response(comprobante)
-        except Exception as e:
-            return Response({'detail': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-class VentaFilter(FilterSet):
-    ven_fecha = DateFromToRangeFilter(field_name='ven_fecha')
-    ven_idcli = NumberFilter(field_name='ven_idcli')
-    ven_idvdo = NumberFilter(field_name='ven_idvdo')
-    ven_estado = NumberFilter(field_name='ven_estado')
-    comprobante = NumberFilter(field_name='comprobante')
-    comprobante_tipo = CharFilter(field_name='comprobante__tipo', lookup_expr='iexact')
-    comprobante_letra = CharFilter(field_name='comprobante__letra', lookup_expr='iexact')
-    ven_sucursal = NumberFilter(field_name='ven_sucursal')
-    ven_total = DateFromToRangeFilter(field_name='ven_total')
-    ven_punto = NumberFilter(field_name='ven_punto')
-    ven_numero = NumberFilter(field_name='ven_numero')
-
-    class Meta:
-        model = Venta
-        fields = [
-            'ven_fecha', 'ven_idcli', 'ven_idvdo', 'ven_estado', 'comprobante',
-            'comprobante_tipo', 'comprobante_letra', 'ven_sucursal', 'ven_total', 'ven_punto', 'ven_numero'
-        ]
-
-class VentaCalculadaFilter(FilterSet):
-    ven_fecha = DateFromToRangeFilter(field_name='ven_fecha')
-    ven_idcli = NumberFilter(field_name='ven_idcli')
-    ven_idvdo = NumberFilter(field_name='ven_idvdo')
-    comprobante_tipo = CharFilter(field_name='comprobante_tipo', lookup_expr='iexact')
-    comprobante_letra = CharFilter(field_name='comprobante_letra', lookup_expr='iexact')
-    # NUEVO: Filtro para notas de crédito - solo facturas válidas
-    para_nota_credito = CharFilter(method='filter_para_nota_credito', label='Para nota de crédito')
-
-    def filter_para_nota_credito(self, queryset, name, value):
-        """
-        Filtra para mostrar solo facturas válidas para asociar a una nota de crédito.
-        Solo permite facturas fiscales (A, B, C) e internas (I).
-        """
-        import logging
-        logger = logging.getLogger(__name__)
-        
-        logger.info(f"[FILTRO] filter_para_nota_credito llamado con value: {value}")
-        logger.info(f"[FILTRO] Queryset inicial: {queryset.count()} registros")
-        
-        if value.lower() == 'true':
-            filtered_queryset = queryset.filter(
-                comprobante_tipo__in=['factura', 'venta', 'factura_interna']
-            )
-            logger.info(f"[FILTRO] Queryset filtrado: {filtered_queryset.count()} registros")
-            
-            # DEBUG: Mostrar tipos encontrados
-            tipos_encontrados = set(filtered_queryset.values_list('comprobante_tipo', flat=True))
-            logger.info(f"[FILTRO] Tipos de comprobantes encontrados: {tipos_encontrados}")
-            
-            return filtered_queryset
-        return queryset
-
-    class Meta:
-        model = VentaCalculada  # Vista SQL (managed = False)
-        fields = ['ven_fecha', 'ven_idcli', 'ven_idvdo', 'comprobante_tipo', 'comprobante_letra', 'para_nota_credito']
-
-class VentaViewSet(viewsets.ModelViewSet):
-    """ViewSet principal para Ventas.
-
-    • list  -> usa la vista calculada (VENTA_CALCULADO) para incluir los totales.
-    • otras -> continúan usando el modelo base `Venta`.
-    """
-
-    # Configuración por defecto (para acciones distintas de list)
-    queryset = Venta.objects.all()
-    serializer_class = VentaSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_class = VentaFilter
-
-    # --- Selección dinámica de queryset / serializer / filterset --------
-    def get_queryset(self):
-        if getattr(self, 'action', None) == 'list':
-            # Aseguramos que el filtro use el modelo adecuado para la vista
-            self.filterset_class = VentaCalculadaFilter
-            # Orden inverso por fecha e ID para traer los más recientes primero
-            return VentaCalculada.objects.all().order_by('-ven_fecha', '-ven_id')
-        # Restablecemos el filtro original para otras acciones
-        self.filterset_class = VentaFilter
-        return super().get_queryset()
-
-    def get_serializer_class(self):
-        if getattr(self, 'action', None) == 'list':
-            return VentaCalculadaSerializer
-        return super().get_serializer_class()
-
-    def get_filterset_class(self):
-        if getattr(self, 'action', None) == 'list':
-            return VentaCalculadaFilter
-        return super().get_filterset_class()
-
-    @transaction.atomic
-    def create(self, request, *args, **kwargs):
-        data = request.data
-        items = data.get('items', [])
-        tipo_comprobante = data.get('tipo_comprobante')
-        # Para Notas de Débito y su equivalente interno permitimos items vacíos;
-        # el serializer generará el ítem genérico en base a los campos específicos.
-        if not items and tipo_comprobante not in ['nota_debito', 'nota_debito_interna']:
-            return Response({'detail': 'El campo items es requerido y no puede estar vacío'}, status=status.HTTP_400_BAD_REQUEST)
-
-        # --- NUEVO: Asignar bonificación general a los ítems sin bonificación particular ---
-        bonif_general = data.get('bonificacionGeneral', 0)
-        try:
-            bonif_general = float(bonif_general)
-        except Exception:
-            bonif_general = 0
-        for item in items:
-            bonif = item.get('vdi_bonifica')
-            if not bonif or float(bonif) == 0:
-                item['vdi_bonifica'] = bonif_general
-
-        # Obtener configuración de la ferretería para determinar política de stock negativo
-        ferreteria = Ferreteria.objects.first()
-        # Usar configuración de la ferretería, con posibilidad de override desde el frontend
-        permitir_stock_negativo = data.get('permitir_stock_negativo', getattr(ferreteria, 'permitir_stock_negativo', False))
-        
-        es_presupuesto = (tipo_comprobante == 'presupuesto')
-        es_nota_credito = (tipo_comprobante == 'nota_credito')
-        es_nota_debito = (tipo_comprobante == 'nota_debito')
-        errores_stock = []
-        stock_actualizado = []
-        if not es_presupuesto:
-            for item in items:
-                id_stock = item.get('vdi_idsto')
-                cantidad = Decimal(str(item.get('vdi_cantidad', 0)))
-
-                # Si el ítem no tiene un ID de stock, es genérico y no participa en la lógica de inventario.
-                if not id_stock:
-                    continue
-
-                # NUEVO: El backend obtiene automáticamente el proveedor habitual del stock
-                # El frontend solo debe enviar vdi_idsto, el backend maneja toda la lógica
-                id_proveedor = _obtener_proveedor_habitual_stock(id_stock)
-                if not id_proveedor:
-                    cod = _obtener_codigo_venta(id_stock)
-                    errores_stock.append(f"No se pudo obtener el proveedor habitual para el producto {cod} (ID: {id_stock})")
-                    continue
-                    
-                if es_nota_credito:
-                    # Para notas de crédito, el stock se devuelve (suma) SOLO al proveedor indicado
-                    try:
-                        stockprove = StockProve.objects.select_for_update().get(stock_id=id_stock, proveedor_id=id_proveedor)
-                    except StockProve.DoesNotExist:
-                        cod = _obtener_codigo_venta(id_stock)
-                        errores_stock.append(f"No existe stock para el producto {cod}")
-                        continue
-                    stockprove.cantidad += cantidad
-                    stockprove.save()
-                    stock_actualizado.append((id_stock, id_proveedor, stockprove.cantidad))
-                # Notas de débito: no tocan stock (no hay ItemsGrid de productos)
-                elif es_nota_debito:
-                    continue
-                    try:
-                        stockprove = StockProve.objects.select_for_update().get(stock_id=id_stock, proveedor_id=id_proveedor)
-                    except StockProve.DoesNotExist:
-                        cod = _obtener_codigo_venta(id_stock)
-                        errores_stock.append(f"No existe stock para el producto {cod}")
-                        continue
-                    stockprove.cantidad += cantidad
-                    stockprove.save()
-                    stock_actualizado.append((id_stock, id_proveedor, stockprove.cantidad))
-                else:
-                    # Venta: descontar distribuyendo entre proveedores si hace falta
-                    _descontar_distribuyendo(
-                        stock_id=id_stock,
-                        proveedor_preferido_id=id_proveedor,
-                        cantidad=cantidad,
-                        permitir_stock_negativo=permitir_stock_negativo,
-                        errores_stock=errores_stock,
-                        stock_actualizado=stock_actualizado,
-                    )
-            if errores_stock:
-                # Imitar exactamente el formato de conversión de presupuesto: detail = str({...})
-                payload = {'detail': 'Error de stock', 'errores': errores_stock}
-                return Response({'detail': str(payload)}, status=status.HTTP_400_BAD_REQUEST)
-        cliente_id = data.get('ven_idcli')
-        cliente = Cliente.objects.filter(id=cliente_id).first()
-        situacion_iva_ferreteria = getattr(ferreteria, 'situacion_iva', None)
-        tipo_iva_cliente = (cliente.iva.nombre if cliente and cliente.iva else '').strip().lower()
-        
-        # Obtener el comprobante apropiado según el tipo y cliente
-        comprobante_id_enviado = data.get('comprobante_id')
-        
-        # Si el frontend envió un comprobante específico, lo usamos (validando que exista)
-        if comprobante_id_enviado:
-            comprobante_obj = Comprobante.objects.filter(codigo_afip=comprobante_id_enviado, activo=True).first()
-            if not comprobante_obj:
-                return Response({
-                    'detail': f'No se encontró comprobante con código AFIP {comprobante_id_enviado} o no está activo'
-                }, status=status.HTTP_400_BAD_REQUEST)
-            comprobante = _construir_respuesta_comprobante(comprobante_obj)
-        else:
-            # Si no se envió un comprobante específico, utilizar la función asignar_comprobante
-            try:
-                comprobante = asignar_comprobante(tipo_comprobante, tipo_iva_cliente)
-            except ValidationError as e:
-                return Response({'detail': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-        
-        if not comprobante:
-            return Response({
-                'detail': 'No se encontró comprobante válido para la operación. '
-                          'Verifique la configuración de comprobantes y letras.'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        data['comprobante_id'] = comprobante["codigo_afip"]
-
-        # === PUNTO DE VENTA PARA COMPROBANTES INTERNOS ===
-        # Si el tipo de comprobante está listado como interno, usar el PV interno 0099
-        if tipo_comprobante in COMPROBANTES_INTERNOS:
-            data['ven_punto'] = PUNTO_VENTA_INTERNO
-
-        # === ALINEAR PUNTO DE VENTA CON ARCA CUANDO ES COMPROBANTE FISCAL ===
-        # Para evitar desalineación con AFIP y colisiones de numeración, si el comprobante
-        # requiere emisión ARCA, forzar a usar el punto de venta configurado en Ferretería.
-        if debe_emitir_arca(tipo_comprobante):
-            pv_arca = getattr(ferreteria, 'punto_venta_arca', None)
-            if pv_arca:
-                data['ven_punto'] = pv_arca
-        # Si el frontend no envió punto de venta, usar el configurado en ferretería como valor por defecto
-        if not data.get('ven_punto'):
-            pv_defecto = getattr(ferreteria, 'punto_venta_arca', None)
-            if pv_defecto:
-                data['ven_punto'] = pv_defecto
-
-        punto_venta = data.get('ven_punto')
-        if not punto_venta:
-            return Response({'detail': 'El punto de venta es requerido'}, status=status.HTTP_400_BAD_REQUEST)
-        intentos = 0
-        max_intentos = 10
-        while intentos < max_intentos:
-            ultima_venta = Venta.objects.filter(
-                ven_punto=punto_venta,
-                comprobante_id=comprobante["codigo_afip"]
-            ).order_by('-ven_numero').first()
-            nuevo_numero = 1 if not ultima_venta else ultima_venta.ven_numero + 1
-            data['ven_numero'] = nuevo_numero
-            try:
-                # === CREAR VENTA ===
-                response = super().create(request, *args, **kwargs)
-                venta_creada = Venta.objects.get(ven_id=response.data['ven_id'])
-                
-                # === INTEGRACIÓN ARCA AUTOMÁTICA (DENTRO DE LA TRANSACCIÓN) ===
-                if debe_emitir_arca(tipo_comprobante):
-                    try:
-                        logger.info(f"Emisión automática ARCA para venta {venta_creada.ven_id} - tipo: {tipo_comprobante}")
-                        resultado_arca = emitir_arca_automatico(venta_creada)
-                        
-                        # Agregar información ARCA a la respuesta
-                        response.data['arca_emitido'] = True
-                        response.data['cae'] = resultado_arca.get('resultado', {}).get('cae')
-                        response.data['cae_vencimiento'] = resultado_arca.get('resultado', {}).get('cae_vencimiento')
-                        response.data['qr_generado'] = resultado_arca.get('resultado', {}).get('qr_generado', False)
-                        response.data['observaciones'] = resultado_arca.get('resultado', {}).get('observaciones', [])
-                        
-                        logger.info(f"Emisión ARCA exitosa para venta {venta_creada.ven_id}: CAE {resultado_arca.get('resultado', {}).get('cae')}")
-                        
-                    except Exception as e:
-                        # Error en emisión ARCA - FALLAR LA TRANSACCIÓN COMPLETA
-                        logger.error(f"Error en emisión automática ARCA para venta {venta_creada.ven_id}: {e}")
-                        raise FerreDeskARCAError(f"Error en emisión ARCA: {e}")
-                else:
-                    # Comprobante interno - no requiere emisión ARCA
-                    response.data['arca_emitido'] = False
-                    response.data['arca_motivo'] = 'Comprobante interno - no requiere emisión ARCA'
-                
-                # === CREAR AUTO-IMPUTACIÓN SI ES "FACTURA RECIBO" ===
-                comprobante_pagado = data.get('comprobante_pagado', False)
-                monto_pago = Decimal(str(data.get('monto_pago', 0)))
-                
-                # Obtener total de la venta desde VentaCalculada
-                from ferreapps.ventas.models import VentaCalculada
-                venta_calculada = VentaCalculada.objects.filter(ven_id=venta_creada.ven_id).first()
-                total_venta = Decimal(str(venta_calculada.ven_total)) if venta_calculada else Decimal('0')
-                
-                # Auto-imputación: solo hasta el total de la venta
-                if comprobante_pagado and monto_pago > 0:
-                    from ferreapps.cuenta_corriente.models import ImputacionVenta
-                    from datetime import date
-                    
-                    # El monto de auto-imputación es el mínimo entre monto_pago y total_venta
-                    monto_auto_imputacion = min(monto_pago, total_venta)
-                    
-                    # Crear la auto-imputación: factura se imputa a sí misma
-                    ImputacionVenta.objects.create(
-                        imp_id_venta=venta_creada,      # La factura
-                        imp_id_recibo=venta_creada,     # La misma factura (auto-imputación)
-                        imp_monto=monto_auto_imputacion, # Monto del pago (limitado al total)
-                        imp_fecha=date.today(),         # Fecha actual
-                        imp_observacion='Factura Recibo - Auto-imputación'
-                    )
-                    
-                    response.data['auto_imputacion_creada'] = True
-                    response.data['monto_imputado'] = str(monto_auto_imputacion)
-                else:
-                    response.data['auto_imputacion_creada'] = False
-                
-                # === CREAR RECIBO DE EXCEDENTE SI EXISTE ===
-                recibo_excedente_data = data.get('recibo_excedente')
-                if recibo_excedente_data:
-                    # Validar que el monto del recibo coincida con el excedente
-                    excedente_calculado = max(monto_pago - total_venta, Decimal('0'))
-                    monto_recibo = Decimal(str(recibo_excedente_data.get('rec_monto_total', 0)))
-                    
-                    if abs(monto_recibo - excedente_calculado) > Decimal('0.01'):  # Tolerancia de 1 centavo
-                        raise ValidationError(
-                            f'El monto del recibo ({monto_recibo}) no coincide con el excedente ({excedente_calculado})'
-                        )
-                    
-                    # Validar que el recibo no tenga imputaciones
-                    if recibo_excedente_data.get('imputaciones'):
-                        raise ValidationError('El recibo de excedente no debe tener imputaciones')
-                    
-                    # Crear el recibo
-                    from ferreapps.ventas.models import VentaDetalleItem
-                    from datetime import date as datetime_date
-                    
-                    # Obtener comprobante de recibo (letra X)
-                    comprobante_recibo = Comprobante.objects.filter(
-                        tipo='recibo',
-                        letra='X',
-                        activo=True
-                    ).first()
-                    
-                    if not comprobante_recibo:
-                        raise ValidationError('No se encontró comprobante de recibo con letra X')
-                    
-                    # Formatear punto de venta y número
-                    rec_pv = int(recibo_excedente_data['rec_pv'])
-                    rec_num = int(recibo_excedente_data['rec_numero'])
-                    
-                    # Verificar unicidad
-                    ya_existe = Venta.objects.filter(
-                        comprobante=comprobante_recibo,
-                        ven_punto=rec_pv,
-                        ven_numero=rec_num
-                    ).exists()
-                    
-                    if ya_existe:
-                        raise ValidationError(
-                            f'El número de recibo X {rec_pv:04d}-{rec_num:08d} ya existe'
-                        )
-                    
-                    # Crear recibo
-                    recibo = Venta.objects.create(
-                        ven_sucursal=1,
-                        ven_fecha=recibo_excedente_data.get('rec_fecha', datetime_date.today()),
-                        comprobante=comprobante_recibo,
-                        ven_punto=rec_pv,
-                        ven_numero=rec_num,
-                        ven_descu1=0,
-                        ven_descu2=0,
-                        ven_descu3=0,
-                        ven_vdocomvta=0,
-                        ven_vdocomcob=0,
-                        ven_estado='CO',
-                        ven_idcli=venta_creada.ven_idcli,
-                        ven_cuit=venta_creada.ven_cuit or '',
-                        ven_dni='',
-                        ven_domicilio=venta_creada.ven_domicilio or '',
-                        ven_razon_social=venta_creada.ven_razon_social or '',
-                        ven_idpla=venta_creada.ven_idpla,
-                        ven_idvdo=venta_creada.ven_idvdo,
-                        ven_copia=1,
-                        ven_observacion=recibo_excedente_data.get('rec_observacion', '')
-                    )
-                    
-                    # Crear item genérico para el recibo
-                    VentaDetalleItem.objects.create(
-                        vdi_idve=recibo,
-                        vdi_idsto=None,
-                        vdi_idpro=None,
-                        vdi_cantidad=1,
-                        vdi_precio_unitario_final=monto_recibo,
-                        vdi_idaliiva=3,  # Alícuota 0%
-                        vdi_orden=1,
-                        vdi_bonifica=0,
-                        vdi_costo=0,
-                        vdi_margen=0,
-                        vdi_detalle1=f'Recibo X {rec_pv:04d}-{rec_num:08d}',
-                        vdi_detalle2=''
-                    )
-                    
-                    response.data['recibo_excedente_creado'] = True
-                    response.data['recibo_excedente_id'] = recibo.ven_id
-                    response.data['recibo_excedente_numero'] = f'X {rec_pv:04d}-{rec_num:08d}'
-                else:
-                    response.data['recibo_excedente_creado'] = False
-                
-                # Agregar datos de respuesta
-                response.data['stock_actualizado'] = stock_actualizado
-                response.data['comprobante_letra'] = comprobante["letra"]
-                response.data['comprobante_nombre'] = comprobante["nombre"]
-                response.data['comprobante_codigo_afip'] = comprobante["codigo_afip"]
-                
-                return response
-            except IntegrityError as e:
-                if 'unique' in str(e).lower() or 'duplicate' in str(e).lower():
-                    intentos += 1
-                    continue
-                else:
-                    raise
-        return Response({'detail': 'No se pudo asignar un número único de venta tras varios intentos.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    @action(detail=True, methods=['post'], url_path='convertir-a-venta')
-    @transaction.atomic
-    def convertir_a_venta(self, request, pk=None):
-        venta = get_object_or_404(Venta, pk=pk)
-        try:
-            if venta.comprobante and (venta.comprobante.tipo == 'presupuesto' or venta.comprobante.nombre.lower().startswith('presupuesto')):
-                items = VentaDetalleItem.objects.filter(vdi_idve=venta.ven_id)
-                # Obtener configuración de la ferretería para determinar política de stock negativo
-                ferreteria = Ferreteria.objects.first()
-                # Usar configuración de la ferretería, con posibilidad de override desde el frontend
-                permitir_stock_negativo = request.data.get('permitir_stock_negativo', getattr(ferreteria, 'permitir_stock_negativo', False))
-                errores_stock = []
-                stock_actualizado = []
-                for item in items:
-                    id_stock = item.vdi_idsto
-                    cantidad = Decimal(str(item.vdi_cantidad))
-                    if not id_stock:
-                        errores_stock.append(f"Falta stock en item: {item.id}")
-                        continue
-                    
-                    # NUEVO: El backend obtiene automáticamente el proveedor habitual del stock
-                    id_proveedor = _obtener_proveedor_habitual_stock(id_stock)
-                    if not id_proveedor:
-                        cod = _obtener_codigo_venta(id_stock)
-                        errores_stock.append(f"No se pudo obtener el proveedor habitual para el producto {cod} (ID: {id_stock})")
-                        continue
-                    
-                    # Descontar distribuyendo entre proveedores si hace falta
-                    _descontar_distribuyendo(
-                        stock_id=id_stock,
-                        proveedor_preferido_id=id_proveedor,
-                        cantidad=cantidad,
-                        permitir_stock_negativo=permitir_stock_negativo,
-                        errores_stock=errores_stock,
-                        stock_actualizado=stock_actualizado,
-                    )
-                if errores_stock:
-                    payload = {'detail': 'Error de stock', 'errores': errores_stock}
-                    return Response({'detail': str(payload)}, status=status.HTTP_400_BAD_REQUEST)
-                cliente = Cliente.objects.filter(id=venta.ven_idcli).first()
-                situacion_iva_ferreteria = getattr(ferreteria, 'situacion_iva', None)
-                tipo_iva_cliente = (cliente.iva.nombre if cliente and cliente.iva else '').strip().lower()
-                comprobante_venta = asignar_comprobante('factura', tipo_iva_cliente)
-                if not comprobante_venta:
-                    return Response({'detail': 'No se encontró comprobante de tipo factura para la conversión.'}, status=status.HTTP_400_BAD_REQUEST)
-                venta.comprobante_id = comprobante_venta["codigo_afip"]
-                venta.ven_estado = 'CE'
-                venta.save()
-                serializer = self.get_serializer(venta)
-                data = serializer.data
-                data['stock_actualizado'] = stock_actualizado
-                data['comprobante_letra'] = comprobante_venta["letra"]
-                data['comprobante_nombre'] = comprobante_venta["nombre"]
-                data['comprobante_codigo_afip'] = comprobante_venta["codigo_afip"]
-                return Response(data)
-            else:
-                return Response({'detail': 'Este documento no es un presupuesto o ya fue convertido.'}, status=status.HTTP_400_BAD_REQUEST)
-        except Exception as e:
-            return Response({'detail': f'Error al convertir: {str(e)}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
-
-    @transaction.atomic
-    def update(self, request, *args, **kwargs):
-        partial = kwargs.pop('partial', False)
-        instance = self.get_object()
-        
-        # ATENCIÓN: Ya no se calculan ni manipulan campos calculados (ven_impneto, ven_total, etc.) aquí.
-        # Toda la lógica de totales y cálculos se delega a la vista SQL.
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-
-        items_data = request.data.get('items', None)
-        if items_data is not None:
-            try:
-                # ATENCIÓN: No calcular totales ni campos calculados aquí.
-                # Solo actualizar los ítems base.
-                # --- NUEVO: Asignar bonificación general a los ítems sin bonificación particular ---
-                bonif_general = request.data.get('bonificacionGeneral', 0)
-                try:
-                    bonif_general = float(bonif_general)
-                except Exception:
-                    bonif_general = 0
-                for item in items_data:
-                    bonif = item.get('vdi_bonifica')
-                    if not bonif or float(bonif) == 0:
-                        item['vdi_bonifica'] = bonif_general
-                # -------------------------------------------------------------------------------
-                instance.items.all().delete()
-                for item_data in items_data:
-                    item_data['vdi_idve'] = instance
-                    for campo_calculado in ['vdi_importe', 'vdi_importe_total', 'vdi_ivaitem']:
-                        item_data.pop(campo_calculado, None)
-                    VentaDetalleItem.objects.create(**item_data)
-            except Exception as e:
-                logging.error(f"Error actualizando ítems de venta: {e}")
-                raise
-
-        return Response(serializer.data)
-
-class VentaDetalleItemViewSet(viewsets.ModelViewSet):
-    queryset = VentaDetalleItem.objects.all()
-    serializer_class = VentaDetalleItemSerializer
-
-class VentaDetalleManViewSet(viewsets.ModelViewSet):
-    queryset = VentaDetalleMan.objects.all()
-    serializer_class = VentaDetalleManSerializer
-
-class VentaRemPedViewSet(viewsets.ModelViewSet):
-    queryset = VentaRemPed.objects.all()
-    serializer_class = VentaRemPedSerializer
-
-class VentaDetalleItemCalculadoFilter(FilterSet):
-    vdi_idve = NumberFilter(field_name='vdi_idve')
-    
-    class Meta:
-        model = VentaDetalleItemCalculado
-        fields = ['vdi_idve']
-
-class VentaDetalleItemCalculadoViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = VentaDetalleItemCalculado.objects.all()
-    serializer_class = VentaDetalleItemCalculadoSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_class = VentaDetalleItemCalculadoFilter
-
-class VentaIVAAlicuotaFilter(FilterSet):
-    vdi_idve = NumberFilter(field_name='vdi_idve')
-    
-    class Meta:
-        model = VentaIVAAlicuota
-        fields = ['vdi_idve']
-
-class VentaIVAAlicuotaViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = VentaIVAAlicuota.objects.all()
-    serializer_class = VentaIVAAlicuotaSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_class = VentaIVAAlicuotaFilter
-
-class VentaCalculadaViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = VentaCalculada.objects.all()
-    serializer_class = VentaCalculadaSerializer
-
-@api_view(['POST'])
-@permission_classes([IsAuthenticated])
-def convertir_presupuesto_a_venta(request):
-    try:
-        with transaction.atomic():
-            print("DEBUG - request.data:", request.data)
-            print("DEBUG - presupuesto_origen:", request.data.get('presupuesto_origen'), type(request.data.get('presupuesto_origen')))
-            print("DEBUG - items_seleccionados:", request.data.get('items_seleccionados'), type(request.data.get('items_seleccionados')))
-            data = request.data
-            presupuesto_id = data.get('presupuesto_origen')
-            items_seleccionados = data.get('items_seleccionados', [])
-            venta_data = data.copy()
-            venta_data.pop('presupuesto_origen', None)
-            venta_data.pop('items_seleccionados', None)
-
-            if not presupuesto_id or not items_seleccionados:
-                error_msg = 'Faltan datos de presupuesto o ítems seleccionados.'
-                if not presupuesto_id:
-                    error_msg += ' No se recibió presupuesto_origen.'
-                if not items_seleccionados:
-                    error_msg += f' items_seleccionados está vacío o None (recibido: {items_seleccionados}).'
-                print(f"DEBUG - Error de validación: {error_msg}")
-                raise ValidationError(error_msg)
-
-            print("DEBUG - INICIO BLOQUE ATOMICO")
-            # Obtener el presupuesto con bloqueo
-            presupuesto = Venta.objects.select_for_update().get(ven_id=presupuesto_id)
-            print("DEBUG - Presupuesto obtenido:", presupuesto)
-            # Validar que es un presupuesto (estado AB)
-            if presupuesto.ven_estado != 'AB':
-                print("DEBUG - Presupuesto no está en estado AB")
-                raise Exception('Solo se pueden convertir presupuestos (estado AB).')
-
-            # Obtener items del presupuesto
-            items_presupuesto = list(presupuesto.items.all())
-            ids_items_presupuesto = [str(item.id) for item in items_presupuesto]
-            print("DEBUG - IDs items presupuesto:", ids_items_presupuesto)
-            # Validar que los ítems seleccionados pertenecen al presupuesto
-            if not all(str(i) in ids_items_presupuesto for i in items_seleccionados):
-                print("DEBUG - Algunos ítems seleccionados no pertenecen al presupuesto")
-                raise Exception('Algunos ítems seleccionados no pertenecen al presupuesto.')
-
-            # === COPIAR ITEMS SELECCIONADOS DEL PRESUPUESTO A venta_data ===
-            # Convertir items del presupuesto al formato que espera el serializer
-            items_para_venta = []
-            for item_presupuesto in items_presupuesto:
-                if str(item_presupuesto.id) in [str(i) for i in items_seleccionados]:
-                    # Convertir el item del presupuesto al formato del serializer
-                    # IMPORTANTE: Los campos son IntegerField, no ForeignKey, por eso no tienen _id
-                    item_data = {
-                        'vdi_idsto': item_presupuesto.vdi_idsto if item_presupuesto.vdi_idsto else None,
-                        'vdi_idpro': item_presupuesto.vdi_idpro if item_presupuesto.vdi_idpro else None,
-                        'vdi_cantidad': float(item_presupuesto.vdi_cantidad),
-                        'vdi_precio_unitario_final': float(item_presupuesto.vdi_precio_unitario_final),
-                        'vdi_idaliiva': item_presupuesto.vdi_idaliiva if item_presupuesto.vdi_idaliiva else None,
-                        'vdi_orden': item_presupuesto.vdi_orden or 1,
-                        'vdi_bonifica': float(item_presupuesto.vdi_bonifica) if item_presupuesto.vdi_bonifica else 0,
-                        'vdi_costo': float(item_presupuesto.vdi_costo) if item_presupuesto.vdi_costo else 0,
-                        'vdi_margen': float(item_presupuesto.vdi_margen) if item_presupuesto.vdi_margen else 0,
-                        'vdi_detalle1': item_presupuesto.vdi_detalle1 or '',
-                        'vdi_detalle2': item_presupuesto.vdi_detalle2 or ''
-                    }
-                    items_para_venta.append(item_data)
-            
-            print(f"DEBUG - Items copiados del presupuesto: {len(items_para_venta)} items")
-            # Agregar los items a venta_data para que el serializer los procese
-            venta_data['items'] = items_para_venta
-
-            # Obtener configuración de la ferretería para determinar política de stock negativo
-            ferreteria = Ferreteria.objects.first()
-            # Usar configuración de la ferretería, con posibilidad de override desde el frontend
-            permitir_stock_negativo = data.get('permitir_stock_negativo', getattr(ferreteria, 'permitir_stock_negativo', False))
-            
-            # Validar stock si es necesario (sumando entre TODOS los proveedores del producto)
-            if not permitir_stock_negativo:
-                errores_stock = []
-                for item in venta_data.get('items', []):
-                    stock_id = item.get('vdi_idsto')
-                    if not stock_id:
-                        continue
-                    try:
-                        cantidad_req = Decimal(str(item.get('vdi_cantidad', 0)))
-                    except Exception:
-                        cantidad_req = Decimal('0')
-                    total_disponible, _ = _total_disponible_en_proveedores(stock_id)
-                    if total_disponible < cantidad_req:
-                        cod = _obtener_codigo_venta(stock_id)
-                        errores_stock.append(
-                            f"Stock insuficiente para producto {cod}. Disponible total: {total_disponible}, solicitado: {cantidad_req}"
-                        )
-                if errores_stock:
-                    print("DEBUG - Errores de stock:", errores_stock)
-                    raise Exception({'detail': 'Error de stock', 'errores': errores_stock})
-
-            # Preparar datos de la venta
-            venta_data['ven_estado'] = 'CE'  # Estado Venta
-            venta_data['ven_tipo'] = 'Venta'  # Tipo Venta
-
-            # --- Lógica de numeración robusta ---
-            tipo_comprobante = venta_data.get('tipo_comprobante')
-            if not tipo_comprobante:
-                print("DEBUG - Falta tipo de comprobante")
-                raise Exception('El tipo de comprobante es requerido')
-
-            # Obtener la situación fiscal de la ferretería y el cliente
-            cliente = Cliente.objects.get(id=venta_data.get('ven_idcli'))
-            situacion_iva_ferreteria = ferreteria.situacion_iva
-            tipo_iva_cliente = cliente.iva.nombre.strip().lower()
-
-            # Usar asignar_comprobante para determinar el comprobante correcto
-            comprobante = asignar_comprobante(tipo_comprobante, tipo_iva_cliente)
-            venta_data['comprobante_id'] = comprobante['codigo_afip']
-
-            # === PUNTO DE VENTA PARA COMPROBANTES INTERNOS ===
-            # Si el tipo de comprobante está listado como interno, usar el PV interno 0099
-            if tipo_comprobante in COMPROBANTES_INTERNOS:
-                venta_data['ven_punto'] = PUNTO_VENTA_INTERNO
-
-            # === ALINEAR PUNTO DE VENTA CON ARCA CUANDO ES COMPROBANTE FISCAL ===
-            # Igual que en el alta: ARCA decide el número final. Para evitar colisiones de la
-            # clave única al renumerar, si es fiscal, usar el PV fiscal configurado.
-            if debe_emitir_arca(tipo_comprobante):
-                pv_arca = getattr(ferreteria, 'punto_venta_arca', None)
-                if pv_arca:
-                    venta_data['ven_punto'] = pv_arca
-            # Si no vino punto de venta (p.ej. factura interna/Cotización), usar PV por defecto
-            if not venta_data.get('ven_punto'):
-                pv_defecto = getattr(ferreteria, 'punto_venta_arca', None)
-                if pv_defecto:
-                    venta_data['ven_punto'] = pv_defecto
-            punto_venta = venta_data.get('ven_punto')
-            if not punto_venta:
-                print("DEBUG - Falta punto de venta")
-                raise Exception('El punto de venta es requerido')
-
-            # --- Asignar bonificación general a los ítems sin bonificación particular ---
-            bonif_general = venta_data.get('bonificacionGeneral', 0)
-            try:
-                bonif_general = float(bonif_general)
-            except Exception:
-                bonif_general = 0
-            for item in venta_data.get('items', []):
-                bonif = item.get('vdi_bonifica')
-                if not bonif or float(bonif) == 0:
-                    item['vdi_bonifica'] = bonif_general
-
-            intentos = 0
-            max_intentos = 10
-            venta = None
-            print("LOG: Antes de crear la venta")
-            while intentos < max_intentos:
-                ultima_venta = Venta.objects.filter(
-                    ven_punto=punto_venta,
-                    comprobante_id=comprobante['codigo_afip']
-                ).order_by('-ven_numero').first()
-                nuevo_numero = 1 if not ultima_venta else ultima_venta.ven_numero + 1
-                venta_data['ven_numero'] = nuevo_numero
-                
-                # === USAR EL MISMO PATRÓN QUE CREATE(): USAR SERIALIZER COMO VENTAFORM ===
-                # Esto permite que Django maneje automáticamente los ForeignKey como ven_idcli
-                try:
-                    # === REPLICAR PATRÓN DE VENTAFORM.CREATE() ===
-                    # 1. Crear venta usando serializer
-                    serializer = VentaSerializer(data=venta_data)
-                    serializer.is_valid(raise_exception=True)
-                    venta = serializer.save()
-                    
-                    # 2. Obtener venta recién creada (igual que VentaForm.create())
-                    venta_creada = Venta.objects.get(ven_id=venta.ven_id)
-                    
-                    print(f"LOG: Venta creada con ID {venta_creada.ven_id}")
-                    
-                    # === CREAR AUTO-IMPUTACIÓN SI ES FACTURA PAGADA ===
-                    comprobante_pagado = venta_data.get('comprobante_pagado', False)
-                    monto_pago = Decimal(str(venta_data.get('monto_pago', 0)))
-                    
-                    if comprobante_pagado and monto_pago > 0:
-                        # Obtener total desde VentaCalculada (vista SQL con campos calculados)
-                        venta_calculada = VentaCalculada.objects.filter(ven_id=venta_creada.ven_id).first()
-                        total_venta = Decimal(str(venta_calculada.ven_total)) if venta_calculada else Decimal('0')
-                        monto_auto_imputacion = min(monto_pago, total_venta)
-                        
-                        from ferreapps.cuenta_corriente.models import ImputacionVenta
-                        from datetime import date
-                        
-                        ImputacionVenta.objects.create(
-                            imp_id_venta=venta_creada,
-                            imp_id_recibo=venta_creada,
-                            imp_monto=monto_auto_imputacion,
-                            imp_fecha=date.today(),
-                            imp_observacion='Factura Recibo - Auto-imputación'
-                        )
-                    
-                    # === CREAR RECIBO DE EXCEDENTE SI EXISTE ===
-                    recibo_excedente_data = data.get('recibo_excedente')
-                    if recibo_excedente_data:
-                        # Obtener total desde VentaCalculada (vista SQL con campos calculados)
-                        venta_calculada = VentaCalculada.objects.filter(ven_id=venta_creada.ven_id).first()
-                        total_venta = Decimal(str(venta_calculada.ven_total)) if venta_calculada else Decimal('0')
-                        excedente_calculado = max(monto_pago - total_venta, Decimal('0'))
-                        monto_recibo = Decimal(str(recibo_excedente_data.get('rec_monto_total', 0)))
-                        
-                        if abs(monto_recibo - excedente_calculado) > Decimal('0.01'):  # Tolerancia de 1 centavo
-                            raise ValidationError(
-                                f'El monto del recibo ({monto_recibo}) no coincide con el excedente ({excedente_calculado})'
-                            )
-                        
-                        # Validar que el recibo no tenga imputaciones
-                        if recibo_excedente_data.get('imputaciones'):
-                            raise ValidationError('El recibo de excedente no debe tener imputaciones')
-                        
-                        # Crear el recibo
-                        from ferreapps.ventas.models import VentaDetalleItem
-                        from datetime import date as datetime_date
-                        
-                        # Obtener comprobante de recibo (letra X)
-                        comprobante_recibo = Comprobante.objects.filter(
-                            tipo='recibo',
-                            letra='X',
-                            activo=True
-                        ).first()
-                        
-                        if not comprobante_recibo:
-                            raise ValidationError('No se encontró comprobante de recibo con letra X')
-                        
-                        # Formatear punto de venta y número
-                        rec_pv = int(recibo_excedente_data['rec_pv'])
-                        rec_num = int(recibo_excedente_data['rec_numero'])
-                        
-                        # Verificar unicidad
-                        ya_existe = Venta.objects.filter(
-                            comprobante=comprobante_recibo,
-                            ven_punto=rec_pv,
-                            ven_numero=rec_num
-                        ).exists()
-                        
-                        if ya_existe:
-                            raise ValidationError(
-                                f'El número de recibo X {rec_pv:04d}-{rec_num:08d} ya existe'
-                            )
-                        
-                        # Crear recibo
-                        recibo = Venta.objects.create(
-                            ven_sucursal=1,
-                            ven_fecha=recibo_excedente_data.get('rec_fecha', datetime_date.today()),
-                            comprobante=comprobante_recibo,
-                            ven_punto=rec_pv,
-                            ven_numero=rec_num,
-                            ven_descu1=0,
-                            ven_descu2=0,
-                            ven_descu3=0,
-                            ven_vdocomvta=0,
-                            ven_vdocomcob=0,
-                            ven_estado='CO',
-                            ven_idcli=venta_creada.ven_idcli,
-                            ven_cuit=venta_creada.ven_cuit or '',
-                            ven_dni='',
-                            ven_domicilio=venta_creada.ven_domicilio or '',
-                            ven_razon_social=venta_creada.ven_razon_social or '',
-                            ven_idpla=venta_creada.ven_idpla,
-                            ven_idvdo=venta_creada.ven_idvdo,
-                            ven_copia=1,
-                            ven_observacion=recibo_excedente_data.get('rec_observacion', '')
-                        )
-                        
-                        # Crear item genérico para el recibo
-                        VentaDetalleItem.objects.create(
-                            vdi_idve=recibo,
-                            vdi_idsto=None,
-                            vdi_idpro=None,
-                            vdi_cantidad=1,
-                            vdi_precio_unitario_final=monto_recibo,
-                            vdi_idaliiva=3,  # Alícuota 0%
-                            vdi_orden=1,
-                            vdi_bonifica=0,
-                            vdi_costo=0,
-                            vdi_margen=0,
-                            vdi_detalle1=f'Recibo X {rec_pv:04d}-{rec_num:08d}',
-                            vdi_detalle2=''
-                        )
-                    
-                    break
-                except IntegrityError as e:
-                    if 'unique' in str(e).lower() or 'duplicate' in str(e).lower():
-                        intentos += 1
-                        continue  # Reintentar con el siguiente número
-                    else:
-                        print("DEBUG - Error de integridad:", e)
-                        raise
-            else:
-                print("DEBUG - No se pudo asignar un número único de venta tras varios intentos.")
-                raise Exception('No se pudo asignar un número único de venta tras varios intentos.')
-
-            print("LOG: Antes de editar/eliminar el presupuesto")
-            # Convertir ambos arrays de IDs a int para comparar correctamente
-            ids_items_presupuesto_int = [int(i.id) for i in items_presupuesto]
-            items_seleccionados_int = [int(i) for i in items_seleccionados]
-            print(f"DEBUG - ids_items_presupuesto_int: {ids_items_presupuesto_int} (type: {type(ids_items_presupuesto_int[0]) if ids_items_presupuesto_int else None})")
-            print(f"DEBUG - items_seleccionados_int: {items_seleccionados_int} (type: {type(items_seleccionados_int[0]) if items_seleccionados_int else None})")
-            if set(items_seleccionados_int) == set(ids_items_presupuesto_int):
-                print("DEBUG - Se seleccionaron todos los ítems, eliminando presupuesto")
-                presupuesto.delete()
-                print("LOG: Presupuesto eliminado")
-                presupuesto_result = None
-            else:
-                print("DEBUG - Se dejan ítems no seleccionados en el presupuesto")
-                items_restantes = [item for item in items_presupuesto if int(item.id) not in items_seleccionados_int]
-                presupuesto.items.set(items_restantes)
-                # Eliminar físicamente los ítems seleccionados del presupuesto
-                ids_a_eliminar = [item.id for item in items_presupuesto if int(item.id) in items_seleccionados_int]
-                if ids_a_eliminar:
-                    VentaDetalleItem.objects.filter(id__in=ids_a_eliminar).delete()
-                # Si no quedan ítems, eliminar el presupuesto
-                if not items_restantes:
-                    print("LOG: Presupuesto quedó vacío tras conversión, eliminando presupuesto")
-                    presupuesto.delete()
-                    presupuesto_result = None
-                else:
-                    # Los totales se calculan automáticamente en las vistas SQL
-                    # No es necesario recalcular campos que no existen en el modelo Venta
-                    presupuesto.save()
-                    presupuesto_result = VentaSerializer(presupuesto).data
-
-            print("LOG: Antes de actualizar stock")
-            # Actualizar stock: aplicar EXACTAMENTE lo mismo que se validó arriba (sobre los items enviados)
-            stock_actualizado = []
-            errores_en_descuento = []
-            for item in venta_data.get('items', []):
-                id_stock_conv = item.get('vdi_idsto')
-                if not id_stock_conv:
-                    continue
-                
-                # NUEVO: El backend obtiene automáticamente el proveedor habitual del stock
-                # El frontend solo debe enviar vdi_idsto, el backend maneja toda la lógica
-                id_prov_conv = _obtener_proveedor_habitual_stock(id_stock_conv)
-                if not id_prov_conv:
-                    cod = _obtener_codigo_venta(id_stock_conv)
-                    errores_en_descuento.append(f"No se pudo obtener el proveedor habitual para el producto {cod} (ID: {id_stock_conv})")
-                    continue
-                
-                cantidad_conv = item.get('vdi_cantidad', 0)
-                
-                ok = _descontar_distribuyendo(
-                    stock_id=id_stock_conv,
-                    proveedor_preferido_id=id_prov_conv,
-                    cantidad=cantidad_conv,
-                    permitir_stock_negativo=venta_data.get('permitir_stock_negativo', getattr(ferreteria, 'permitir_stock_negativo', False)),
-                    errores_stock=errores_en_descuento,
-                    stock_actualizado=stock_actualizado,
-                )
-                if not ok:
-                    payload = {'detail': 'Error de stock', 'errores': errores_en_descuento}
-                    raise Exception(str(payload))
-
-            # === INTEGRACIÓN ARCA AUTOMÁTICA (DENTRO DE LA TRANSACCIÓN) ===
-            if debe_emitir_arca(tipo_comprobante):
-                try:
-                    logger.info(f"Emisión automática ARCA para conversión presupuesto {presupuesto.ven_id} a venta {venta.ven_id} - tipo: {tipo_comprobante}")
-                    resultado_arca = emitir_arca_automatico(venta)
-                    
-                    # Agregar información ARCA a la respuesta
-                    response_data = {
-                        'venta': VentaSerializer(venta).data,
-                        'presupuesto': presupuesto_result,
-                        'stock_actualizado': stock_actualizado,
-                        'comprobante_letra': comprobante['letra'],
-                        'comprobante_nombre': comprobante['nombre'],
-                        'comprobante_codigo_afip': comprobante['codigo_afip'],
-                        'arca_emitido': True,
-                        'cae': resultado_arca.get('resultado', {}).get('cae'),
-                        'cae_vencimiento': resultado_arca.get('resultado', {}).get('cae_vencimiento'),
-                        'qr_generado': resultado_arca.get('resultado', {}).get('qr_generado', False),
-                        'observaciones': resultado_arca.get('resultado', {}).get('observaciones', [])
-                    }
-                    
-                    logger.info(f"Emisión ARCA exitosa para conversión presupuesto {presupuesto.ven_id} a venta {venta.ven_id}: CAE {resultado_arca.get('resultado', {}).get('cae')}")
-                    
-                except Exception as e:
-                    # Error en emisión ARCA - FALLAR LA TRANSACCIÓN COMPLETA
-                    logger.error(f"Error en emisión automática ARCA para conversión presupuesto {presupuesto.ven_id} a venta {venta.ven_id}: {e}")
-                    raise FerreDeskARCAError(f"Error en emisión ARCA: {e}")
-            else:
-                # Comprobante interno - no requiere emisión ARCA
-                response_data = {
-                    'venta': VentaSerializer(venta).data,
-                    'presupuesto': presupuesto_result,
-                    'stock_actualizado': stock_actualizado,
-                    'comprobante_letra': comprobante['letra'],
-                    'comprobante_nombre': comprobante['nombre'],
-                    'comprobante_codigo_afip': comprobante['codigo_afip'],
-                    'arca_emitido': False,
-                    'arca_motivo': 'Comprobante interno - no requiere emisión ARCA'
-                }
-            
-            return Response(response_data)
-
-    except Exception as e:
-        print("DEBUG - Error en convertir_presupuesto_a_venta:", str(e))
-        return Response({'detail': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-
-
-
-@api_view(['POST'])
-@permission_classes([IsAuthenticated])
-@transaction.atomic
-def convertir_factura_interna_a_fiscal(request):
-    """
-    Convierte una factura interna a factura fiscal.
-    Diferencia clave: items originales NO descontan stock nuevamente.
-    """
-    try:
-        data = request.data
-        factura_interna_id = data.get('factura_interna_origen')
-        tipo_conversion = data.get('tipo_conversion')
-        
-        # Validar tipo de conversión
-        if tipo_conversion != 'factura_i_factura':
-            return Response({'detail': 'Tipo de conversión inválido'}, status=status.HTTP_400_BAD_REQUEST)
-        
-        # Obtener factura interna original
-        try:
-            factura_interna = Venta.objects.select_for_update().get(
-                ven_id=factura_interna_id,
-                comprobante__tipo='factura_interna'
-            )
-        except Venta.DoesNotExist:
-            return Response({'detail': 'Factura interna no encontrada'}, status=status.HTTP_404_NOT_FOUND)
-        
-        # Preparar datos de la nueva factura fiscal (IDÉNTICO A VENTAFORM)
-        venta_data = data.copy()
-        venta_data.pop('factura_interna_origen', None)
-        venta_data.pop('tipo_conversion', None)
-        venta_data.pop('conversion_metadata', None)
-        
-        venta_data['ven_estado'] = 'CE'  # Estado Cerrado para factura fiscal
-        
-        # === USAR EL MISMO FLUJO QUE VentaViewSet.create() ===
-        
-        # Para conversión de factura interna, obtener items originales y marcarlos como no descuenten stock
-        from ferreapps.ventas.models import VentaDetalleItem
-        items_originales = VentaDetalleItem.objects.filter(vdi_idve=factura_interna)
-        
-        # Convertir items originales al formato esperado por el serializer, marcándolos como originales
-        items = []
-        for item_original in items_originales:
-            item_data = {
-                'vdi_orden': item_original.vdi_orden,
-                'vdi_idsto': item_original.vdi_idsto,
-                'vdi_idpro': item_original.vdi_idpro,
-                'vdi_cantidad': item_original.vdi_cantidad,
-                'vdi_costo': item_original.vdi_costo,
-                'vdi_margen': item_original.vdi_margen,
-                'vdi_precio_unitario_final': item_original.vdi_precio_unitario_final,
-                'vdi_bonifica': item_original.vdi_bonifica,
-                'vdi_detalle1': item_original.vdi_detalle1,
-                'vdi_detalle2': item_original.vdi_detalle2,
-                'vdi_idaliiva': item_original.vdi_idaliiva,
-                # Marcar como item original para que no descuente stock
-                'idOriginal': item_original.id,
-                'noDescontarStock': True,
-                'esBloqueado': True,
-            }
-            items.append(item_data)
-        
-        print(f"LOG: Obtenidos {len(items)} items originales de la factura interna {factura_interna.ven_id}")
-
-        # === LÓGICA DIFERENCIADA DE STOCK PARA CONVERSIONES ===
-        # Obtener configuración de la ferretería para determinar política de stock negativo
-        ferreteria = Ferreteria.objects.first()
-        # Usar configuración de la ferretería, con posibilidad de override desde el frontend
-        permitir_stock_negativo = venta_data.get('permitir_stock_negativo', getattr(ferreteria, 'permitir_stock_negativo', False))
-        
-        tipo_comprobante = venta_data.get('tipo_comprobante')
-        es_presupuesto = (tipo_comprobante == 'presupuesto')
-        es_nota_credito = (tipo_comprobante == 'nota_credito')
-        errores_stock = []
-        stock_actualizado = []
-        
-        if not es_presupuesto:
-            for item in items:
-                id_stock = item.get('vdi_idsto')
-                cantidad = Decimal(str(item.get('vdi_cantidad', 0)))
-
-                # Si el ítem no tiene un ID de stock, es genérico y no participa en la lógica de inventario.
-                if not id_stock:
-                    continue
-
-                # NUEVO: Si tiene idOriginal, noDescontarStock o esBloqueado, proviene de factura interna - NO descontar stock
-                if item.get('idOriginal') or item.get('noDescontarStock') or item.get('esBloqueado'):
-                    print(f"LOG: Item original/noDescontarStock/esBloqueado - NO descuenta stock (idOriginal={item.get('idOriginal')}, noDescontarStock={item.get('noDescontarStock')}, esBloqueado={item.get('esBloqueado')})")
-                    continue
-
-                # Si es un producto real NUEVO, el backend obtiene automáticamente el proveedor habitual
-                id_proveedor = _obtener_proveedor_habitual_stock(id_stock)
-                if not id_proveedor:
-                    cod = _obtener_codigo_venta(id_stock)
-                    errores_stock.append(f"No se pudo obtener el proveedor habitual para el producto {cod} (ID: {id_stock})")
-                    continue
-                    
-                if es_nota_credito:
-                    # Para notas de crédito, el stock se devuelve (suma) SOLO al proveedor indicado
-                    try:
-                        stockprove = StockProve.objects.select_for_update().get(stock_id=id_stock, proveedor_id=id_proveedor)
-                    except StockProve.DoesNotExist:
-                        cod = _obtener_codigo_venta(id_stock)
-                        errores_stock.append(f"No existe stock para el producto {cod}")
-                        continue
-                    stockprove.cantidad += cantidad
-                    stockprove.save()
-                    stock_actualizado.append((id_stock, id_proveedor, stockprove.cantidad))
-                else:
-                    # Para ventas normales, descontar distribuyendo entre proveedores si hace falta
-                    _descontar_distribuyendo(
-                        stock_id=id_stock,
-                        proveedor_preferido_id=id_proveedor,
-                        cantidad=cantidad,
-                        permitir_stock_negativo=permitir_stock_negativo,
-                        errores_stock=errores_stock,
-                        stock_actualizado=stock_actualizado,
-                    )
-                
-            if errores_stock:
-                payload = {'detail': 'Error de stock', 'errores': errores_stock}
-                return Response({'detail': str(payload)}, status=status.HTTP_400_BAD_REQUEST)
-
-        # === LÓGICA DE COMPROBANTE (IDÉNTICA AL MÉTODO CREATE) ===
-        # Obtener ferretería para leer configuración fiscal y PV
-        ferreteria = Ferreteria.objects.first()
-        cliente_id = venta_data.get('ven_idcli')
-        cliente = Cliente.objects.filter(id=cliente_id).first()
-        situacion_iva_ferreteria = getattr(ferreteria, 'situacion_iva', None)
-        tipo_iva_cliente = (cliente.iva.nombre if cliente and cliente.iva else '').strip().lower()
-        
-        # Obtener el comprobante apropiado según el tipo y cliente
-        comprobante_id_enviado = venta_data.get('comprobante_id')
-        
-        # Si el frontend envió un comprobante específico, lo usamos (validando que exista)
-        if comprobante_id_enviado:
-            comprobante_obj = Comprobante.objects.filter(codigo_afip=comprobante_id_enviado, activo=True).first()
-            if not comprobante_obj:
-                return Response({
-                    'detail': f'No se encontró comprobante con código AFIP {comprobante_id_enviado} o no está activo'
-                }, status=status.HTTP_400_BAD_REQUEST)
-            comprobante = _construir_respuesta_comprobante(comprobante_obj)
-        else:
-            # Si no se envió un comprobante específico, utilizar la función asignar_comprobante
-            try:
-                comprobante = asignar_comprobante(tipo_comprobante, tipo_iva_cliente)
-            except ValidationError as e:
-                return Response({'detail': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-        
-        if not comprobante:
-            return Response({
-                'detail': 'No se encontró comprobante válido para la operación. '
-                          'Verifique la configuración de comprobantes y letras.'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        venta_data['comprobante_id'] = comprobante["codigo_afip"]
-
-        # === ALINEAR PUNTO DE VENTA CON ARCA CUANDO ES COMPROBANTE FISCAL ===
-        # Mismo criterio que en create: ARCA define el número final. Para evitar colisiones
-        # de la clave única (ven_punto, ven_numero, comprobante), si el comprobante requiere
-        # emisión ARCA, forzamos a usar el punto de venta fiscal configurado en ferretería
-        # antes de asignar numeración provisional.
-        tipo_comprobante = venta_data.get('tipo_comprobante')
-        if debe_emitir_arca(tipo_comprobante):
-            pv_arca = getattr(ferreteria, 'punto_venta_arca', None)
-            if pv_arca:
-                venta_data['ven_punto'] = pv_arca
-
-        # === LÓGICA DE NUMERACIÓN (IDÉNTICA AL MÉTODO CREATE) ===
-        punto_venta = venta_data.get('ven_punto')
-        if not punto_venta:
-            return Response({'detail': 'El punto de venta es requerido'}, status=status.HTTP_400_BAD_REQUEST)
-        
-        intentos = 0
-        max_intentos = 10
-        nueva_factura = None
-        
-        while intentos < max_intentos:
-            ultima_venta = Venta.objects.filter(
-                ven_punto=punto_venta,
-                comprobante_id=comprobante["codigo_afip"]
-            ).order_by('-ven_numero').first()
-            nuevo_numero = 1 if not ultima_venta else ultima_venta.ven_numero + 1
-            venta_data['ven_numero'] = nuevo_numero
-            
-            # Agregar items originales al venta_data para el serializer
-            venta_data['items'] = items
-            
-            # Limpiar flags personalizados de los items antes de pasarlos al serializer
-            # El serializer no reconoce estos campos y pueden causar errores
-            for item in venta_data['items']:
-                item.pop('idOriginal', None)
-                item.pop('noDescontarStock', None)
-                item.pop('esBloqueado', None)
-            
-            # === USAR EL MISMO PATRÓN QUE CREATE(): USAR SERIALIZER COMO VENTAFORM ===
-            # Esto permite que Django maneje automáticamente los ForeignKey como ven_idcli
-            try:
-                # === REPLICAR PATRÓN DE VENTAFORM.CREATE() ===
-                # 1. Crear venta usando serializer
-                serializer = VentaSerializer(data=venta_data)
-                serializer.is_valid(raise_exception=True)
-                nueva_factura = serializer.save()
-                
-                # 2. Obtener venta recién creada (igual que VentaForm.create())
-                nueva_factura = Venta.objects.get(ven_id=nueva_factura.ven_id)
-                
-                print(f"LOG: Factura fiscal creada con ID {nueva_factura.ven_id}")
-                
-                # === CREAR AUTO-IMPUTACIÓN SI ES FACTURA PAGADA ===
-                comprobante_pagado = venta_data.get('comprobante_pagado', False)
-                monto_pago = Decimal(str(venta_data.get('monto_pago', 0)))
-                
-                if comprobante_pagado and monto_pago > 0:
-                    # Obtener total desde VentaCalculada (vista SQL con campos calculados)
-                    venta_calculada = VentaCalculada.objects.filter(ven_id=nueva_factura.ven_id).first()
-                    total_venta = Decimal(str(venta_calculada.ven_total)) if venta_calculada else Decimal('0')
-                    monto_auto_imputacion = min(monto_pago, total_venta)
-                    
-                    from ferreapps.cuenta_corriente.models import ImputacionVenta
-                    from datetime import date
-                    
-                    ImputacionVenta.objects.create(
-                        imp_id_venta=nueva_factura,
-                        imp_id_recibo=nueva_factura,
-                        imp_monto=monto_auto_imputacion,
-                        imp_fecha=date.today(),
-                        imp_observacion='Factura Recibo - Auto-imputación'
-                    )
-                
-                # === CREAR RECIBO DE EXCEDENTE SI EXISTE ===
-                recibo_excedente_data = data.get('recibo_excedente')
-                if recibo_excedente_data:
-                    # Obtener total desde VentaCalculada (vista SQL con campos calculados)
-                    venta_calculada = VentaCalculada.objects.filter(ven_id=nueva_factura.ven_id).first()
-                    total_venta = Decimal(str(venta_calculada.ven_total)) if venta_calculada else Decimal('0')
-                    excedente_calculado = max(monto_pago - total_venta, Decimal('0'))
-                    monto_recibo = Decimal(str(recibo_excedente_data.get('rec_monto_total', 0)))
-                    
-                    if abs(monto_recibo - excedente_calculado) > Decimal('0.01'):  # Tolerancia de 1 centavo
-                        raise ValidationError(
-                            f'El monto del recibo ({monto_recibo}) no coincide con el excedente ({excedente_calculado})'
-                        )
-                    
-                    # Validar que el recibo no tenga imputaciones
-                    if recibo_excedente_data.get('imputaciones'):
-                        raise ValidationError('El recibo de excedente no debe tener imputaciones')
-                    
-                    # Crear el recibo
-                    from ferreapps.ventas.models import VentaDetalleItem
-                    from datetime import date as datetime_date
-                    
-                    # Obtener comprobante de recibo (letra X)
-                    comprobante_recibo = Comprobante.objects.filter(
-                        tipo='recibo',
-                        letra='X',
-                        activo=True
-                    ).first()
-                    
-                    if not comprobante_recibo:
-                        raise ValidationError('No se encontró comprobante de recibo con letra X')
-                    
-                    # Formatear punto de venta y número
-                    rec_pv = int(recibo_excedente_data['rec_pv'])
-                    rec_num = int(recibo_excedente_data['rec_numero'])
-                    
-                    # Verificar unicidad
-                    ya_existe = Venta.objects.filter(
-                        comprobante=comprobante_recibo,
-                        ven_punto=rec_pv,
-                        ven_numero=rec_num
-                    ).exists()
-                    
-                    if ya_existe:
-                        raise ValidationError(
-                            f'El número de recibo X {rec_pv:04d}-{rec_num:08d} ya existe'
-                        )
-                    
-                    # Crear recibo
-                    recibo = Venta.objects.create(
-                        ven_sucursal=1,
-                        ven_fecha=recibo_excedente_data.get('rec_fecha', datetime_date.today()),
-                        comprobante=comprobante_recibo,
-                        ven_punto=rec_pv,
-                        ven_numero=rec_num,
-                        ven_descu1=0,
-                        ven_descu2=0,
-                        ven_descu3=0,
-                        ven_vdocomvta=0,
-                        ven_vdocomcob=0,
-                        ven_estado='CO',
-                        ven_idcli=nueva_factura.ven_idcli,
-                        ven_cuit=nueva_factura.ven_cuit or '',
-                        ven_dni='',
-                        ven_domicilio=nueva_factura.ven_domicilio or '',
-                        ven_razon_social=nueva_factura.ven_razon_social or '',
-                        ven_idpla=nueva_factura.ven_idpla,
-                        ven_idvdo=nueva_factura.ven_idvdo,
-                        ven_copia=1,
-                        ven_observacion=recibo_excedente_data.get('rec_observacion', '')
-                    )
-                    
-                    # Crear item genérico para el recibo
-                    VentaDetalleItem.objects.create(
-                        vdi_idve=recibo,
-                        vdi_idsto=None,
-                        vdi_idpro=None,
-                        vdi_cantidad=1,
-                        vdi_precio_unitario_final=monto_recibo,
-                        vdi_idaliiva=3,  # Alícuota 0%
-                        vdi_orden=1,
-                        vdi_bonifica=0,
-                        vdi_costo=0,
-                        vdi_margen=0,
-                        vdi_detalle1=f'Recibo X {rec_pv:04d}-{rec_num:08d}',
-                        vdi_detalle2=''
-                    )
-                
-                break
-            except IntegrityError as e:
-                if 'unique' in str(e).lower() or 'duplicate' in str(e).lower():
-                    intentos += 1
-                    continue  # Reintentar con el siguiente número
-                else:
-                    print("DEBUG - Error de integridad:", e)
-                    raise
-        else:
-            print("DEBUG - No se pudo asignar un número único de factura tras varios intentos.")
-            return Response({'detail': 'No se pudo asignar un número único de factura tras varios intentos.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        
-        print("LOG: Antes de gestionar la factura interna original")
-        
-        # === LÓGICA DE GESTIÓN DE FACTURA INTERNA ===
-        # Conversión siempre completa: eliminar la factura interna original
-        try:
-            print("DEBUG - Eliminando factura interna original")
-            factura_interna.delete()
-            print("LOG: Factura interna eliminada exitosamente")
-            factura_interna_result = None
-        except ProtectedError as e:
-            # Si tiene imputaciones, capturar el error y dar contexto al usuario
-            from ferreapps.cuenta_corriente.models import ImputacionVenta
-            from django.db.models import Q
-            
-            print(f"LOG: No se puede eliminar factura interna por imputaciones: {e}")
-            
-            # Obtener todas las imputaciones relacionadas
-            imputaciones_relacionadas = ImputacionVenta.objects.filter(
-                Q(imp_id_venta=factura_interna) | Q(imp_id_recibo=factura_interna)
-            ).select_related('imp_id_venta', 'imp_id_recibo')
-            
-            # Construir lista de comprobantes relacionados
-            comprobantes_relacionados = []
-            for imp in imputaciones_relacionadas:
-                # Validar que no sea auto-imputación (ej. factura-recibo)
-                if imp.imp_id_venta.ven_id == imp.imp_id_recibo.ven_id:
-                    # Auto-imputación: la cotización se pagó a sí misma (factura-recibo)
-                    vc = VentaCalculada.objects.filter(ven_id=imp.imp_id_venta.ven_id).first()
-                    if vc:
-                        comprobantes_relacionados.append({
-                            'tipo': 'auto_imputacion',
-                            'numero': vc.numero_formateado,
-                            'nombre': vc.comprobante_nombre + ' (Factura-Recibo)',
-                            'monto': str(imp.imp_monto),
-                            'fecha': str(imp.imp_fecha)
-                        })
-                    continue
-                
-                if imp.imp_id_venta.ven_id == factura_interna.ven_id:
-                    # Esta cotización está siendo pagada por otro comprobante
-                    vc = VentaCalculada.objects.filter(ven_id=imp.imp_id_recibo.ven_id).first()
-                    if vc:
-                        comprobantes_relacionados.append({
-                            'tipo': 'recibo_pago',
-                            'numero': vc.numero_formateado,
-                            'nombre': vc.comprobante_nombre,
-                            'monto': str(imp.imp_monto),
-                            'fecha': str(imp.imp_fecha)
-                        })
-                elif imp.imp_id_recibo.ven_id == factura_interna.ven_id:
-                    # Esta cotización está pagando otra factura
-                    vc = VentaCalculada.objects.filter(ven_id=imp.imp_id_venta.ven_id).first()
-                    if vc:
-                        comprobantes_relacionados.append({
-                            'tipo': 'factura_pagada',
-                            'numero': vc.numero_formateado,
-                            'nombre': vc.comprobante_nombre,
-                            'monto': str(imp.imp_monto),
-                            'fecha': str(imp.imp_fecha)
-                        })
-            
-            return Response({
-                'detail': 'No se puede convertir esta cotización porque tiene imputaciones realizadas.',
-                'razon': 'La cotización tiene pagos registrados que deben eliminarse antes de la conversión.',
-                'imputaciones': comprobantes_relacionados,
-                'total_imputaciones': len(comprobantes_relacionados)
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        # === INTEGRACIÓN ARCA AUTOMÁTICA (DENTRO DE LA TRANSACCIÓN) ===
-        if debe_emitir_arca(tipo_comprobante):
-            try:
-                logger.info(f"Emisión automática ARCA para conversión factura interna {factura_interna_id} a factura fiscal {nueva_factura.ven_id} - tipo: {tipo_comprobante}")
-                resultado_arca = emitir_arca_automatico(nueva_factura)
-                
-                # Agregar información ARCA a la respuesta
-                response_data = VentaSerializer(nueva_factura).data
-                response_data['stock_actualizado'] = stock_actualizado
-                response_data['comprobante_letra'] = comprobante["letra"]
-                response_data['comprobante_nombre'] = comprobante["nombre"]
-                response_data['comprobante_codigo_afip'] = comprobante["codigo_afip"]
-                response_data['factura_interna'] = factura_interna_result
-                response_data['arca_emitido'] = True
-                response_data['cae'] = resultado_arca.get('resultado', {}).get('cae')
-                response_data['cae_vencimiento'] = resultado_arca.get('resultado', {}).get('cae_vencimiento')
-                response_data['qr_generado'] = resultado_arca.get('resultado', {}).get('qr_generado', False)
-                response_data['observaciones'] = resultado_arca.get('resultado', {}).get('observaciones', [])
-                
-                logger.info(f"Emisión ARCA exitosa para conversión factura interna {factura_interna_id} a factura fiscal {nueva_factura.ven_id}: CAE {resultado_arca.get('resultado', {}).get('cae')}")
-                
-            except Exception as e:
-                # Error en emisión ARCA - FALLAR LA TRANSACCIÓN COMPLETA
-                logger.error(f"Error en emisión automática ARCA para conversión factura interna {factura_interna_id} a factura fiscal {nueva_factura.ven_id}: {e}")
-                raise FerreDeskARCAError(f"Error en emisión ARCA: {e}")
-        else:
-            # Comprobante interno - no requiere emisión ARCA
-            response_data = VentaSerializer(nueva_factura).data
-            response_data['stock_actualizado'] = stock_actualizado
-            response_data['comprobante_letra'] = comprobante["letra"]
-            response_data['comprobante_nombre'] = comprobante["nombre"]
-            response_data['comprobante_codigo_afip'] = comprobante["codigo_afip"]
-            response_data['factura_interna'] = factura_interna_result
-            response_data['arca_emitido'] = False
-            response_data['arca_motivo'] = 'Comprobante interno - no requiere emisión ARCA'
-        
-        return Response(response_data, status=status.HTTP_201_CREATED)
-        
-    except Exception as e:
-        # Replicar comportamiento de create: dejar propagar la excepción para que
-        # @transaction.atomic realice el rollback completo y el handler global
-        # formatee el mensaje hacia el frontend.
-        print("DEBUG - Error en convertir_factura_interna_a_fiscal:", str(e))
-        raise
-
-
-
-# Dashboard endpoints
-from django.http import JsonResponse
-from django.db import connection
-from django.utils import timezone
-from datetime import datetime, timedelta
-import json
-
-def productos_mas_vendidos(request):
-    """Endpoint para obtener los productos más vendidos"""
-    tipo = request.GET.get('tipo', 'cantidad')  # 'cantidad' o 'total'
-    
-    try:
-        with connection.cursor() as cursor:
-            if tipo == 'cantidad':
-                # Agrupar por producto y sumar cantidades
-                query = """
-                SELECT 
-                    COALESCE(vdi.vdi_detalle1, 'Producto sin nombre') as producto,
-                    SUM(vdi.vdi_cantidad) as total_cantidad
-                FROM "VENTADETALLEITEM_CALCULADO" vdi
-                WHERE vdi.vdi_cantidad > 0
-                GROUP BY vdi.vdi_detalle1
-                ORDER BY total_cantidad DESC
-                LIMIT 10
-                """
-            else:
-                # Agrupar por producto y sumar totales facturados
-                query = """
-                SELECT 
-                    COALESCE(vdi.vdi_detalle1, 'Producto sin nombre') as producto,
-                    SUM(vdi.total_item) as total_facturado
-                FROM "VENTADETALLEITEM_CALCULADO" vdi
-                WHERE vdi.total_item > 0
-                GROUP BY vdi.vdi_detalle1
-                ORDER BY total_facturado DESC
-                LIMIT 10
-                """
-            
-            cursor.execute(query)
-            results = cursor.fetchall()
-            
-            if not results:
-                return JsonResponse({
-                    'labels': [],
-                    'datasets': [{
-                        'label': 'Cantidad Vendida' if tipo == 'cantidad' else 'Total Facturado ($)',
-                        'data': [],
-                        'backgroundColor': 'rgba(59, 130, 246, 0.8)',
-                        'borderColor': 'rgba(59, 130, 246, 1)',
-                        'borderWidth': 1,
-                        'borderRadius': 4,
-                    }]
-                })
-            
-            labels = [row[0] for row in results]
-            data = [float(row[1]) for row in results]
-            
-            return JsonResponse({
-                'labels': labels,
-                'datasets': [{
-                    'label': 'Cantidad Vendida' if tipo == 'cantidad' else 'Total Facturado ($)',
-                    'data': data,
-                    'backgroundColor': 'rgba(59, 130, 246, 0.8)',
-                    'borderColor': 'rgba(59, 130, 246, 1)',
-                    'borderWidth': 1,
-                    'borderRadius': 4,
-                }]
-            })
-            
-    except Exception as e:
-        return JsonResponse({'error': str(e)}, status=500)
-
-def ventas_por_dia(request):
-    """Endpoint para obtener las ventas por día"""
-    periodo = request.GET.get('periodo', '7d')
-    
-    # Calcular fechas según el período
-    hoy = timezone.now().date()
-    if periodo == '7d':
-        fecha_inicio = hoy - timedelta(days=7)
-    elif periodo == '30d':
-        fecha_inicio = hoy - timedelta(days=30)
-    elif periodo == '90d':
-        fecha_inicio = hoy - timedelta(days=90)
-    elif periodo == '1y':
-        fecha_inicio = hoy - timedelta(days=365)
-    else:
-        fecha_inicio = hoy - timedelta(days=7)
-    
-    try:
-        with connection.cursor() as cursor:
-            query = """
-            SELECT 
-                DATE(vc.ven_fecha) as fecha,
-                SUM(vc.ven_total) as total_ventas
-            FROM "VENTA_CALCULADO" vc
-            WHERE vc.ven_fecha >= %s AND vc.ven_fecha <= %s
-            GROUP BY DATE(vc.ven_fecha)
-            ORDER BY fecha
-            """
-            
-            cursor.execute(query, [fecha_inicio, hoy])
-            results = cursor.fetchall()
-            
-            if not results:
-                return JsonResponse({
-                    'labels': [],
-                    'datasets': [{
-                        'label': 'Ventas Diarias ($)',
-                        'data': [],
-                        'borderColor': 'rgba(34, 197, 94, 1)',
-                        'backgroundColor': 'rgba(34, 197, 94, 0.1)',
-                        'borderWidth': 3,
-                        'fill': True,
-                        'tension': 0.4,
-                        'pointBackgroundColor': 'rgba(34, 197, 94, 1)',
-                        'pointBorderColor': '#ffffff',
-                        'pointBorderWidth': 2,
-                        'pointRadius': 6,
-                        'pointHoverRadius': 8,
-                    }]
-                })
-            
-            # Procesar resultados reales
-            fechas = []
-            ventas = []
-            for row in results:
-                fecha = row[0]
-                if isinstance(fecha, str):
-                    fecha = datetime.strptime(fecha, '%Y-%m-%d').date()
-                fechas.append(fecha.strftime('%d/%m'))
-                ventas.append(float(row[1]))
-            
-            return JsonResponse({
-                'labels': fechas,
-                'datasets': [{
-                    'label': 'Ventas Diarias ($)',
-                    'data': ventas,
-                    'borderColor': 'rgba(34, 197, 94, 1)',
-                    'backgroundColor': 'rgba(34, 197, 94, 0.1)',
-                    'borderWidth': 3,
-                    'fill': True,
-                    'tension': 0.4,
-                    'pointBackgroundColor': 'rgba(34, 197, 94, 1)',
-                    'pointBorderColor': '#ffffff',
-                    'pointBorderWidth': 2,
-                    'pointRadius': 6,
-                    'pointHoverRadius': 8,
-                }]
-            })
-            
-    except Exception as e:
-        return JsonResponse({'error': str(e)}, status=500)
-
-def clientes_mas_ventas(request):
-    """Endpoint para obtener los clientes con más ventas"""
-    tipo = request.GET.get('tipo', 'total')  # 'total', 'cantidad', 'frecuencia'
-    
-    try:
-        with connection.cursor() as cursor:
-            if tipo == 'total':
-                # Agrupar por cliente y sumar totales facturados
-                query = """
-                SELECT 
-                    COALESCE(vc.cliente_razon, 'Cliente sin nombre') as cliente,
-                    SUM(vc.ven_total) as total_facturado
-                FROM "VENTA_CALCULADO" vc
-                WHERE vc.ven_total > 0
-                GROUP BY vc.cliente_razon
-                ORDER BY total_facturado DESC
-                LIMIT 10
-                """
-            elif tipo == 'cantidad':
-                # Agrupar por cliente y sumar cantidad de productos
-                query = """
-                SELECT 
-                    COALESCE(vc.cliente_razon, 'Cliente sin nombre') as cliente,
-                    SUM(vdi.vdi_cantidad) as total_productos
-                FROM "VENTADETALLEITEM_CALCULADO" vdi
-                JOIN "VENTA_CALCULADO" vc ON vdi.vdi_idve = vc.ven_id
-                WHERE vdi.vdi_cantidad > 0
-                GROUP BY vc.cliente_razon
-                ORDER BY total_productos DESC
-                LIMIT 10
-                """
-            else:  # frecuencia
-                # Agrupar por cliente y contar número de compras
-                query = """
-                SELECT 
-                    COALESCE(vc.cliente_razon, 'Cliente sin nombre') as cliente,
-                    COUNT(DISTINCT vc.ven_id) as frecuencia_compras
-                FROM "VENTA_CALCULADO" vc
-                WHERE vc.ven_total > 0
-                GROUP BY vc.cliente_razon
-                ORDER BY frecuencia_compras DESC
-                LIMIT 10
-                """
-            
-            cursor.execute(query)
-            results = cursor.fetchall()
-            
-            if not results:
-                return JsonResponse({
-                    'labels': [],
-                    'datasets': [{
-                        'label': 'Total Facturado ($)' if tipo == 'total' else 'Cantidad de Productos' if tipo == 'cantidad' else 'Frecuencia de Compras',
-                        'data': [],
-                        'backgroundColor': 'rgba(168, 85, 247, 0.8)',
-                        'borderColor': 'rgba(168, 85, 247, 1)',
-                        'borderWidth': 1,
-                        'borderRadius': 4,
-                    }]
-                })
-            
-            labels = [row[0] for row in results]
-            data = [float(row[1]) for row in results]
-            
-            return JsonResponse({
-                'labels': labels,
-                'datasets': [{
-                    'label': 'Total Facturado ($)' if tipo == 'total' else 'Cantidad de Productos' if tipo == 'cantidad' else 'Frecuencia de Compras',
-                    'data': data,
-                    'backgroundColor': 'rgba(168, 85, 247, 0.8)',
-                    'borderColor': 'rgba(168, 85, 247, 1)',
-                    'borderWidth': 1,
-                    'borderRadius': 4,
-                }]
-            })
-            
-    except Exception as e:
-        return JsonResponse({'error': str(e)}, status=500)
-
-
+# Advertencia para desarrolladores
+warnings.warn(
+    "El archivo views.py está deprecado. "
+    "El código se ha modularizado en ferreapps.ventas.views/",
+    DeprecationWarning,
+    stacklevel=2
+)
