+++ ferredesk_v0/frontend/src/components/Presupuestos y Ventas/hooks/useComprobantesCRUD.js
@@ -426,33 +426,140 @@ const useComprobantesCRUD = ({
     setIsFetchingForConversion(true);
 
     try {
-      // Validación OPCIONAL: verificar imputaciones antes de abrir formulario
-      try {
-        const ccResponse = await fetch(`/api/cuenta-corriente/cliente/${facturaInterna.cliente?.id || facturaInterna.ven_idcli}/`)
-        if (ccResponse.ok) {
-          const ccData = await ccResponse.json()
-          const tieneImputaciones = ccData.items?.some(item => 
-            item.ven_id === facturaInterna.id && item.saldo_pendiente < item.ven_total
-          )
+      // === NUEVA VALIDACIÓN: Verificar imputaciones ANTES de abrir modal ===
+      const validacionResponse = await fetch(`/api/verificar-imputaciones/${facturaInterna.id}/`);
+      const validacionData = await validacionResponse.json();
+      
+      if (validacionData.tiene_imputaciones && !validacionData.puede_convertir) {
+        
+        // CASO ESPECIAL: Cliente genérico con auto-imputaciones (requiere confirmación)
+        if (validacionData.requiere_confirmacion && validacionData.es_cliente_generico && validacionData.puede_eliminar_auto) {
+          // Construir mensaje de confirmación específico
+          let mensajeConfirmacion = `${validacionData.detail}\n\n`;
+          mensajeConfirmacion += `Pagos registrados:\n\n`;
           
-          if (tieneImputaciones) {
-            const confirmar = window.confirm(
-              'ADVERTENCIA: Esta cotización tiene imputaciones en cuenta corriente.\n\n' +
-              'No podrá completar la conversión hasta que elimine las imputaciones.\n\n' +
-              '¿Desea continuar de todos modos para ver los detalles?'
-            )
-            if (!confirmar) {
-              setIsFetchingForConversion(false);
-              setFetchingPresupuestoId(null);
+          validacionData.imputaciones.forEach((imp, idx) => {
+            if (imp.tipo === 'auto_imputacion') {
+              mensajeConfirmacion += `${idx + 1}. ${imp.nombre} ${imp.numero}\n`;
+              mensajeConfirmacion += `   Monto: $${imp.monto} - Fecha: ${imp.fecha}\n\n`;
+            }
+          });
+          
+          mensajeConfirmacion += `Total: ${validacionData.total_imputaciones} pago(s) registrado(s)\n\n`;
+          mensajeConfirmacion += `${validacionData.mensaje_confirmacion}\n\n`;
+          mensajeConfirmacion += `NOTA: Esta opción solo está disponible para ventas del cliente Consumidor Final.`;
+          
+          
+          // Mostrar confirmación
+          const confirmarEliminacion = window.confirm(mensajeConfirmacion);
+          
+          if (!confirmarEliminacion) {
+            // Usuario canceló: no hacer nada
+            return;
+          }
+          
+          // Usuario confirmó: ELIMINAR AUTO-IMPUTACIONES INMEDIATAMENTE
+          try {
+            const responseEliminar = await fetch(
+              `/api/eliminar-auto-imputaciones-cliente-generico/${facturaInterna.id}/`,
+              {
+                method: 'POST',
+                headers: {
+                  'Content-Type': 'application/json',
+                },
+              }
+            );
+            
+            const dataEliminar = await responseEliminar.json();
+            
+            if (!dataEliminar.success) {
+              window.alert(`Error al eliminar auto-imputaciones: ${dataEliminar.detail}`);
               return;
             }
-          }
-        }
-      } catch (ccError) {
-        console.warn('No se pudo verificar imputaciones:', ccError);
-        // Continuar sin validación preventiva
-      }
+            
+          console.log(`Auto-imputaciones eliminadas: ${dataEliminar.num_eliminadas}`);
+          
+          // ✅ Obtener datos y abrir formulario inmediatamente después de eliminar
+          const [cabecera, itemsDetalle] = await Promise.all([
+            fetch(`/api/venta-calculada/${facturaInterna.id}/`).then(async (res) => {
+              if (!res.ok) {
+                const err = await res.json().catch(() => ({ detail: "Error cabecera" }));
+                throw new Error(err.detail);
+              }
+              return res.json();
+            }),
+            fetch(`/api/venta-detalle-item-calculado/?vdi_idve=${facturaInterna.id}`).then(async (res) => {
+              if (!res.ok) return [];
+              return res.json();
+            }),
+          ]);
 
+          const facturaInternaConDetalle = {
+            ...(cabecera.venta || facturaInterna),
+            items: Array.isArray(itemsDetalle) ? itemsDetalle : [],
+          };
+
+          const itemsConId = facturaInternaConDetalle.items.map((it, idx) => ({
+            ...it,
+            id: it.id || it.vdi_idve || it.vdi_id || idx + 1,
+          }));
+
+          const key = `conversion-factura-i-${facturaInterna.id}`
+          const label = `Convertir ${facturaInterna.numero || facturaInterna.id}`
+
+          const tabData = {
+            facturaInternaOrigen: {
+              ...facturaInternaConDetalle, 
+              items: itemsConId,
+              tipoConversion: 'factura_i_factura',
+            },
+            itemsSeleccionados: itemsConId,
+            itemsSeleccionadosIds: itemsConId.map(item => item.id),
+            tipoConversion: 'factura_i_factura'
+          }
+          
+          updateTabData(key, label, tabData, 'conv-factura-i');
+          return; // Salir después de abrir el formulario
+          
+        } catch (error) {
+          window.alert(`Error al eliminar auto-imputaciones: ${error.message}`);
+          return;
+        }
+        } else {
+          // CASO GENERAL: Imputaciones que no se pueden eliminar automáticamente
+          let mensaje = `${validacionData.detail}\n\n`;
+          mensaje += `Comprobantes relacionados:\n\n`;
+          
+          validacionData.imputaciones.forEach((imp, idx) => {
+            if (imp.tipo === 'auto_imputacion') {
+              mensaje += `${idx + 1}. ${imp.nombre} ${imp.numero}\n`;
+              mensaje += `   Tipo: Auto-imputación (Factura-Recibo)\n`;
+              mensaje += `   Monto: $${imp.monto}\n`;
+              mensaje += `   Fecha: ${imp.fecha}\n\n`;
+            } else if (imp.tipo === 'recibo_pago') {
+              mensaje += `${idx + 1}. ${imp.nombre} ${imp.numero}\n`;
+              mensaje += `   Tipo: Pago recibido\n`;
+              mensaje += `   Monto: $${imp.monto}\n`;
+              mensaje += `   Fecha: ${imp.fecha}\n\n`;
+            } else if (imp.tipo === 'factura_pagada') {
+              mensaje += `${idx + 1}. ${imp.nombre} ${imp.numero}\n`;
+              mensaje += `   Tipo: Pago realizado\n`;
+              mensaje += `   Monto: $${imp.monto}\n`;
+              mensaje += `   Fecha: ${imp.fecha}\n\n`;
+            }
+          });
+          
+          mensaje += `Total de imputaciones: ${validacionData.total_imputaciones}\n\n`;
+          mensaje += 'Para continuar, elimine estas imputaciones desde Cuenta Corriente antes de realizar la conversión.';
+          
+          // Mostrar error con alert del navegador (no abrir formulario)
+          window.alert(mensaje);
+          return;
+        }
+      }
+      // === FIN VALIDACIÓN ===
+
+      // Para casos SIN imputaciones: obtener datos y abrir formulario
       const [cabecera, itemsDetalle] = await Promise.all([
         fetch(`/api/venta-calculada/${facturaInterna.id}/`).then(async (res) => {
           if (!res.ok) {
@@ -480,12 +587,12 @@ const useComprobantesCRUD = ({
       // Abrir ConVentaForm directamente (SIN ConversionModal)
       const key = `conversion-factura-i-${facturaInterna.id}`
       const label = `Convertir ${facturaInterna.numero || facturaInterna.id}`
-      
+
       const tabData = {
         facturaInternaOrigen: {
           ...facturaInternaConDetalle, 
           items: itemsConId,
-          tipoConversion: 'factura_i_factura'
+          tipoConversion: 'factura_i_factura',
         },
         itemsSeleccionados: itemsConId,
         itemsSeleccionadosIds: itemsConId.map(item => item.id),
@@ -495,7 +602,7 @@ const useComprobantesCRUD = ({
       updateTabData(key, label, tabData, 'conv-factura-i');
     } catch (error) {
       console.error("Error al obtener detalle para conversión:", error);
-      alert(error.message);
+      window.alert(error.message);
     } finally {
       setIsFetchingForConversion(false);
       setFetchingPresupuestoId(null);
